+++ frontend/src/index.js
@@ -3,6 +3,10 @@ import ReactDOM from 'react-dom'
 import App from './App'
 import MainPage from './pages/MainPage'
 import SingleNote from './pages/notes/[id]'
+import SingleProduct from './pages/products/[id]'
+import CartPage from './pages/CartPage'
+import LoginPage from './pages/LoginPage/LoginPage'
+
 import 'bootstrap/dist/css/bootstrap.min.css'
 import { BudgetsProvider } from './context/BudgetsContext'
 import { NotesProvider } from './context/NotesContext'
@@ -24,7 +28,10 @@ ReactDOM.render(
             <Routes>
               <Route path='/' element={<App />}></Route>
               <Route path='/test' element={<MainPage />}></Route>
+              <Route path='/login' element={<LoginPage />}></Route>
+              <Route path='/cart' element={<CartPage />}></Route>
               <Route path='/notes/:id' element={<SingleNote />}></Route>
+              <Route path='/products/:id' element={<SingleProduct />}></Route>
             </Routes>
           </BrowserRouter>
         </BudgetsProvider>